requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
openai==1.54.5
packaging==24.2
proto-plus==1.25.0
protobuf==5.28.3
psycopg==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wikipedia==1.4.0
